update tests to use new code layout
